refactor:댓글 드롭메뉴 수정
