+++ src/components/Community/CommentDropDown.jsx
@@ -5,6 +5,7 @@ import MoreVertIcon from "@mui/icons-material/MoreVert";
 import api from "../../apis/axios";
 import { useEffect, useState } from "react";
 import { useUserInfo } from "../../store/useUserInfo";
+import { AlarmDialog } from "../../styles/AlarmDialog";
 
 const ITEM_HEIGHT = 48;
 
@@ -24,10 +25,11 @@ export default function CommentDropDown({
   };
 
   const user = useUserInfo((state) => state.user);
-  console.log(user);
+  //console.log(user);
   const myInfo = user?.userId;
+  const isMyComment = myInfo === comment?.userId ? "mycomment" : "notmycomment";
   // post가 없을 경우 기본 UI만 렌더링
-  if (!comment) {
+  if (!comment ) {
     return (
       <IconButton
         aria-label="more"
@@ -44,7 +46,7 @@ export default function CommentDropDown({
   if (!myInfo) {
     return <div>Loading user info...</div>;
   }
-  const isMyComment = myInfo === comment?.userId ? "mycomment" : "notmycomment";
+  
 
   // 옵션 구성 함수
   const setOptions = () => {
@@ -55,12 +57,7 @@ export default function CommentDropDown({
           { label: "수정하기", action: "edit" },
           { label: "삭제하기", action: "delete" },
         ];
-      // 자신이 작성하지 않은 댓글 -> 신고하기, 저장하기
-      case "notmycomment":
-        return [
-          { label: "신고하기", action: "report" },
-          { label: "저장하기", action: "save" },
-        ];
+      // 자신이 작성하지 않은 댓글 -> 메뉴바 보이지않도록
       default:
         return [];
     }
@@ -73,18 +70,19 @@ export default function CommentDropDown({
     if (window.confirm("정말로 이 댓글을 삭제하시겠습니까?")) {
       try {
         const response = await api.delete(`/posts/comments/${comment.id}`);
-        const message = response.data?.message || "댓글이 삭제되었습니다.";
-        console.log(message);
+        response.data?.message || "댓글이 삭제되었습니다.";
+        //console.log(message);
         alert("댓글이 삭제되었습니다.");
         //댓글 리랜더링
         reloadComments();
       } catch (error) {
         const errorMessage = error.response?.data?.message || "댓글 삭제 오류";
-        console.log(errorMessage, error);
+        //console.log(errorMessage, error);
         alert(errorMessage);
       }
     }
   };
+
 
   // 각 옵션별 수행 함수
   const handleAction = (action) => {
@@ -97,12 +95,7 @@ export default function CommentDropDown({
       case "delete":
         deleteComment();
         break;
-      //   // 게시물 신고
-      //   case "report":
-      //     reportPost();
-      //     break;
-      default:
-        break;
+      
     }
   };
