+++ src/styles/AlarmDialog.jsx
